Change naming; Add Camera cases;
